process_requests
